Refactoring: The 'properties' rule becomes 'order' and deals with different kinds of sorting (using option 'statics-first' means that all statics should be declared first). Wip. Changed the tests and documentation
