+++ README.md
@@ -83,48 +83,79 @@ class Foo {
 }
 ```
 
-### properties
+### order
 
-This feature is ES7 ([proposal stage](https://gist.github.com/jeffmo/054df782c05639da2adb)). 
-It has been tested with the non-default parser *babel-eslint*, because at the moment *esprima* doesn't recognize 
+This rule sets the order of members in classes. Basically the properties should be declared before methods.
+
+You can activate the option `'statics-first'`, which says that all statics are declared in the beginning.
+The order would then be: static props, static methods, properties, methods.
+
+The object property feature is ES7 ([proposal stage](https://gist.github.com/jeffmo/054df782c05639da2adb)).
+Note that the tests work with the non-default parser *babel-eslint*, because at the moment *esprima* doesn't recognize
 properties and throws errors.
 
-The following patterns are NOT valid:
+The following patterns are **INVALID**:
 
-```js
-class Foo {
-  baz() {}
-  static bar = 1;
-}
-class Foo {
-  baz() {}
-  bar = 2;
-}  
-```
+1. Without `statics-first`:
+  ```js
+  class Foo {
+    static alpha = 1
+    static gamma() {};
+    beta = 2
+    delta() {}
+  }
 
-The following patterns are valid:
+  class Foo {
+    beta = 2
+    static alpha = 1
+    static gamma() {};
+    delta() {}
+  }
 
-```js
-class Foo {
-  bar = 3; 
-}          
-class Foo {
-  bar = 2;
-  baz() {}
-}        
-class Foo {
-  /* comments allowed */
-  bar = 2;
-  baz() {}
-}        
-class Foo {
-  bar() {}
-}        
-class Foo {
-  static bar = 1;
-  baz() {}
-}
-```
+  class Foo {
+    static alpha = 1
+    static beta() {};
+    gamma = 2
+    delta() {}
+    gamma2 = 3
+  }
+  ```
+2. With `statics-first`:
+  ```js
+  class Foo {
+    static alpha = 1
+    gamma = 2
+    static beta() {};
+    delta() {}
+  }
+  ```
+
+The following patterns are **VALID**:
+
+1. Without `statics-first`:
+  ```js
+  class Foo {
+    static alpha = 1
+    beta = 2
+    static gamma() {};
+    delta() {}
+  }
+
+  class Foo {
+    /* comments, of course, are allowed */
+    beta = 2;
+    delta() {}
+  }
+  ```
+2. With `statics-first`:
+  ```js
+  class Foo {
+    static alpha = 1
+    static beta() {};
+    gamma = 2
+    delta() {}
+  }
+  ```
 
 
 ## Usage
@@ -137,7 +168,7 @@ rules:
   # Plugins
   classes/space  : 2
   classes/name   : [2, "class", "name-required", "method"]
-  classes/properties   : [2, "on-top"]
+  classes/order   : [2, "statics-first"]
 ```
 
 ## License
