+++ index.js
@@ -2,6 +2,6 @@ module.exports = {
   rules: {
     'space': require('./lib/rules/space'),
     'name': require('./lib/rules/name'),
-    'properties': require('./lib/rules/properties')
+    'order': require('./lib/rules/order')
   }
 };
