+++ lib/rules/order.js
@@ -2,6 +2,13 @@
 
 module.exports = function(context) {
 
+  var member = {
+    staticProps: 1,
+    staticMethods: 2,
+    props: 3,
+    methods: 4
+  };
+
   function propertiesOnTop(node) {
 
     if (context.options.indexOf('on-top') < 0) {
@@ -21,7 +28,7 @@ module.exports = function(context) {
           // Weird enough, 'static' is not regarded as a keyword but as class property (at least by babel-eslint)
           // Shouldn't report the error twice, though
           if (crtToken.value !== 'static') {
-            context.report(node, 'properties should be declared on top.', {identifier: crtToken.value});
+            context.report(node, 'properties should be declared berfore methods.', {identifier: crtToken.value});
             break;
           }
         }
